fix API delete supply_product: cannot delete product while it in pending purchaseOrder

+++ server/src/modules/suppliers/suppliers.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateSupplierWithProductsDto } from './dto/create-supplier.dto';
 import { UpdateSupplierDto } from './dto/update-supplier.dto';
 import { PrismaService } from '../prisma/prisma.service';
@@ -206,6 +206,17 @@ export class SuppliersService {
 
     if (!existing) {
       throw new NotFoundException(`Không tìm thấy supply_products với productid = ${productid} và supplierid = ${supplierid}`);
+    }
+
+    const hasPendingOrder = await this.prisma.purchase_order.findFirst({
+      where: {
+        productid,
+        statusorder: 'pending',
+      },
+    });
+
+    if (hasPendingOrder) {
+      throw new BadRequestException(`Không thể xóa sản phẩm vì đang nằm trong đơn hàng ${hasPendingOrder.poid}`);
     }
 
     await this.prisma.supply_products.delete({
